% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{abo.gp}
\alias{abo.gp}
\title{Find the log probability of the data using ABO CV}
\usage{
abo.gp(X, k = makeLaplacianKernel(1), lambda = 0, processMean = 0,
  scl = T, Y = NULL, display = F)
}
\arguments{
\item{X}{Input data, or input and target variables, with Y the last column}

\item{k}{The kernel function}

\item{lambda}{The lambda parameter}

\item{processMean}{The process mean of the models}

\item{scl}{Whether the data should be scaled and centered}

\item{Y}{The target variables, if not included in X. If NULL it is assumed they are included in X.}
}
\value{
The sum of the log probability of each data point using ABO CV
}
\description{
Find the log probability of the data using ABO CV
}
\keyword{internal}
