% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{createKernelMatrix}
\alias{createKernelMatrix}
\title{Create a kernel matrix}
\usage{
createKernelMatrix(x, k)
}
\arguments{
\item{x}{The data points to use}

\item{k}{The kernel function to use}
}
\value{
The kernel matrix
}
\description{
Create a kernel matrix from k(x,x)
}
\keyword{internal}
