% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{createKernelMatrix2}
\alias{createKernelMatrix2}
\title{Create a kernel matrix 2}
\usage{
createKernelMatrix2(x, X, k)
}
\arguments{
\item{x}{The first set of data points to use}

\item{X}{The second set of data points to use}

\item{k}{The kernel function to use}
}
\value{
The kernel matrix
}
\description{
Create a kernel matrix from k(x,X)
}
\keyword{internal}
