% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{emptyProbability}
\alias{emptyProbability}
\title{Probability of Y data using an empty GP}
\usage{
emptyProbability(X, k = makeLaplacianKernel(1), lambda = 0,
  processMean = 0, scl = T, Y = NULL)
}
\arguments{
\item{X}{The X data}

\item{k}{The kernal function}

\item{lambda}{The lambda parameter}

\item{processMean}{The process mean}

\item{scl}{Whether the data should be scaled and centered}

\item{Y}{The Y data as a vector. If NULL, the last column of X is used as Y.}
}
\value{
A scalar giving the probability of the data
}
\description{
Probability of Y data using an empty GP
}
\keyword{internal}
