% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit1}
\alias{fit1}
\title{Fit a single hyper-parameter kernel's parameter and lambda parameter}
\usage{
fit1(X, lambda, sigma, kernelFunc = makeLaplacianKernel, folds = 0,
  processMean = 0, scl = T, Y = NULL, display = F)
}
\arguments{
\item{X}{Input data, or input and target variables, with Y the last column}

\item{lambda}{A vector of lambda parameters to search over}

\item{sigma}{A vector of sigma kernel parameters to search over}

\item{folds}{The number of folds to perform. If 0, all but one validation is performed.}

\item{processMean}{The process mean of the models}

\item{scl}{Whether the data should be scaled and centered}

\item{Y}{The target variables, if not included in X. If NULL it is assumed they are included in X.}
}
\value{
The cross validation likelihood grid for the parameter combinations.
}
\description{
Fit a single hyper-parameter kernel's parameter and lambda parameter
}
