% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Create a Gaussian Process}
\usage{
gp(X, k = makeLaplacianKernel(1), lambda = 0, processMean = 0, scl = T,
  Y = NULL)
}
\arguments{
\item{X}{The X data as a matrix, or able to be case to a matrix}

\item{k}{The kernel function}

\item{lambda}{The lambda parameter}

\item{processMean}{The mean of the process. If NA the mean of the Y data is used.}

\item{scl}{Whether the data should be scaled and centered}

\item{Y}{The Y data as a vector. If NULL, the last column of X is used as Y.}
}
\value{
A gp object
}
\description{
Create a Gaussian Process
}
