% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{gpc}
\alias{gpc}
\title{Produce a logistic GP classifier}
\usage{
gpc(X, Y = NULL, samples = 1000, burn = samples/10, method = "MG",
  kernel = "Laplacian", kernelParams = NULL, fDenom = ifelse(method ==
  "MG", 10, length(Y) + 1), sDenom = 10, fmean = 0, progress = F,
  createKernel = makeLaplacianKernel)
}
\arguments{
\item{X}{The X data as a matrix, or able to be case to a matrix}

\item{Y}{The Y data as a vector. If NULL, the last column of X is used as Y.}

\item{samples}{The number of samples to use}

\item{burn}{The burn period}

\item{method}{The MCMC method to use ("M" is Metropolis; "MG" is Metrolopolis in Gibbs)}

\item{fmean}{The mean of the process for the latent variable. If NA the mean of the 
estimate points is used.}

\item{progress}{Whether progress should be printed}

\item{denom}{Denominator for the standard Gaussian random walk in the Metropolis steps}
}
\value{
A gpc object
}
\description{
Produce a logistic GP classifier.
Uses a Monte Carlo (Metropolis Hastings) Engine.
}
