% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gridSearch1}
\alias{gridSearch1}
\title{Perform a grid search to fit a single hyper-parameter kernel's parameter and lambda parameter}
\usage{
gridSearch1(X, kernelFunc = makeLaplacianKernel, maxIter = 100,
  sensitivity = 0.001, folds = 0, processMean = 0, scl = T, Y = NULL,
  display = F, detailed = F, sigmaMax = 1000, lambdaMax = 1000,
  sigmaMin = 1e-04, lambdaMin = 1e-04)
}
\arguments{
\item{X}{Input data, or input and target variables, with Y the last column}

\item{maxIter}{The maximum number of iterations to continue the search for}

\item{sensitivity}{Required convergence difference to terminate on}

\item{folds}{The number of folds to perform. If 0, all but one validation is performed.}

\item{processMean}{The process mean of the models}

\item{scl}{Whether the data should be scaled and centered}

\item{Y}{The target variables, if not included in X. If NULL it is assumed they are included in X.}
}
\value{
The model built from the best found hyper-parameters
}
\description{
Perform a grid search to fit a single hyper-parameter kernel's parameter and lambda parameter
}
