% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{kfolds.gp}
\alias{kfolds.gp}
\title{Perform K-folds fitting on GP}
\usage{
kfolds.gp(X, folds = 10, k = makeLaplacianKernel(1), lambda = 0,
  processMean = 0, scl = T, Y = NULL)
}
\arguments{
\item{X}{Input data, or input and target variables, with Y the last column}

\item{folds}{The number of folds to use}

\item{k}{The kernel function}

\item{lambda}{The lambda parameter}

\item{processMean}{The process mean of the models}

\item{scl}{Whether the data should be scaled and centered}

\item{Y}{The target variables, if not included in X. If NULL it is assumed they are included in X.}
}
\value{
The sum of the log probability of each data point using K-folds CV
}
\description{
Perform K-folds fitting on GP
}
\keyword{internal}
