% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{makeExponentialKernel}
\alias{makeExponentialKernel}
\title{Make Exponential Kernel}
\usage{
makeExponentialKernel(sigma)
}
\arguments{
\item{sigma}{The sigma parameter}
}
\value{
The specified kernel function
}
\description{
Constructor function for the exponential kernel
}
