% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{makeKernelPacket}
\alias{makeKernelPacket}
\title{Make Kernel Package for GPC}
\usage{
makeKernelPacket(type, params, X)
}
\arguments{
\item{type}{Kernel type: "Exponential", "Laplacian"}
}
\value{
Kernel package for gpc function
}
\description{
Make Kernel Package for GPC
}
