% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{makeLaplacianKernel}
\alias{makeLaplacianKernel}
\title{Make Laplacian Kernel}
\usage{
makeLaplacianKernel(sigma)
}
\arguments{
\item{sigma}{The sigma parameter}
}
\value{
The specified kernel function
}
\description{
Constructor function for the Laplacian kernel
}
