% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gp}
\alias{plot.gp}
\title{Plot Gaussian Process}
\usage{
\method{plot}{gp}(model, ci = 2, steps = 50, top = T, bot = T,
  samples = 0, residuals = T, xLim = NULL, yLim = NULL, ...)
}
\arguments{
\item{model}{The Gaussian process to model}

\item{ci}{The confidence interval level, in terms of standard deviation}

\item{steps}{The number of points to calculate}

\item{top}{Whether the top confidence interval should be shown}

\item{bot}{Whether the bottom confidence interval should be shown}

\item{samples}{The number of function samples to display}

\item{...}{Other arguments to plot or rgl::plot3d}
}
\value{
Invisible NULL
}
\description{
Plot Gaussian Process
}
