% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gpc}
\alias{plot.gpc}
\title{Plot a categorical Gaussian Process}
\usage{
\method{plot}{gpc}(model, ...)
}
\arguments{
\item{model}{The gpc model}

\item{...}{Other arguments to plot or rgl::plot3d}

\item{steps}{The number of points to calculate}

\item{ci}{The confidence interval level, in terms of standard deviation}

\item{top}{Whether the top confidence interval should be shown}

\item{bot}{Whether the bottom confidence interval should be shown}
}
\value{
Invisible NULL
}
\description{
Plot a categorical Gaussian Process
}
