% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gp}
\alias{predict.gp}
\title{Predict}
\usage{
\method{predict}{gp}(model, x, full = quicker(model, x))
}
\arguments{
\item{model}{The gp object}

\item{x}{The X data, as a matrix. Vectors will be interpreted as single row matrices.}

\item{full}{Should the full covariance matrix be used.}
}
\value{
A matrix, with the first column giving predictions (means) and the second
column giving variance.
}
\description{
Predict
}
