% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gpc}
\alias{predict.gpc}
\title{Prediction function for classification model}
\usage{
\method{predict}{gpc}(model, X, ...)
}
\arguments{
\item{model}{gpc model}

\item{X}{Input data}

\item{...}{Other arguments, which are all ignored.}
}
\value{
A three column matrix consisting of the probabilitys predicted
}
\description{
S3 override prediction function for classification model
}
