% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{probability}
\alias{probability}
\title{Find the probability of points}
\usage{
probability(model, X, Y, logprob = T)
}
\arguments{
\item{model}{The gp model}

\item{X}{The input data}

\item{Y}{The target data}

\item{logprob}{Whether the probabilities should be log}
}
\value{
A scalar giving the probability or log probability
}
\description{
Find the probability of points
}
\keyword{internal}
