% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{updateLaplacianKernel}
\alias{updateLaplacianKernel}
\title{Perform a Metropolis step on the Laplacian kernel}
\usage{
updateLaplacianKernel(kPacket, X, Y)
}
\description{
Perform a Metropolis step on the Laplacian kernel
}
